move PublicClassMapper to our package
